pyproject.toml
[tool.poetry]
name = "open-poen-api"
version = "0.1.0"
description = "The API of Open Poen - an open spending platform where we make publicly available how money from the community is spent."
authors = ["Mark de Wijk <dewijkmark@gmail.com>"]
license = "EUPL-1.2"
readme = "README.md"
packages = [{include = "open_poen_api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
httpx = "^0.24.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^1.10.9"}
sqlalchemy-utils = "^0.41.1"
pytz = "^2023.3"
python-dateutil = "^2.8.2"
pycryptodome = "^3.18.0"
nordigen = "^1.3.1"
typer = {extras = ["all"], version = "^0.9.0"}
fastapi-users = {extras = ["sqlalchemy"], version = "^12.0.0"}
asyncpg = "^0.27.0"
fastapi-mail = "^1.3.1"
pytest-asyncio = "^0.21.0"
pytest = "^7.4.0"
oso = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
requests = "^2.31.0"
types-python-dateutil = "^2.8.19.13"
types-pytz = "^2023.3.0.0"
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
asyncio_mode="auto"

[tool.poetry.scripts]
open-poen = "open_poen_api.cli:app"